+++ src/scanner.rs
@@ -1,8 +1,18 @@
-use std::{ fmt::Debug, path::Path, sync::{ Arc, Mutex } };
+use std::{fmt::Debug, path::Path};
 
 use log::trace;
 
-use crate::{ FindMatches, Result, ScannerImpl, ScannerMode };
+use crate::{FindMatches, Result, ScannerImpl, ScannerMode};
+
+/// A trait to switch between scanner modes.
+pub trait ScannerModeSwitcher {
+    /// Sets the current scanner mode.
+    fn set_mode(&mut self, mode: usize);
+    /// Returns the current scanner mode.
+    fn current_mode(&self) -> usize;
+    /// Returns the name of the scanner mode with the given index.
+    fn mode_name(&self, index: usize) -> Option<&str>;
+}
 
 /// A Scanner.
 /// It consists of multiple DFAs that are used to search for matches.
@@ -19,7 +29,6 @@ use crate::{ FindMatches, Result, ScannerImpl, ScannerMode };
 /// At least one scanner mode must be added to the scanner. This is usually the mode named `INITIAL`.
 pub struct Scanner {
     pub(crate) inner: ScannerImpl,
-    current_mode: Arc<Mutex<usize>>,
 }
 
 impl Scanner {
@@ -28,41 +37,16 @@ impl Scanner {
     /// The ScannerImpl is created from the scanner modes and the current mode is shared between
     /// the scanner and the scanner impl.
     pub fn try_new(scanner_modes: Vec<ScannerMode>) -> Result<Self> {
-        let current_mode = Arc::new(Mutex::new(0));
-        let mut me = Scanner {
-            inner: ScannerImpl::try_from(scanner_modes)?,
-            current_mode: current_mode.clone(),
-        };
         // Share the current mode between the scanner and the scanner impl.
-        me.inner.common_mode(current_mode.clone());
-        Ok(me)
+        Ok(Scanner {
+            inner: ScannerImpl::try_from(scanner_modes)?,
+        })
     }
 
     /// Returns an iterator over all non-overlapping matches.
     /// The iterator yields a [`Match`] value until no more matches could be found.
     pub fn find_iter<'h>(&self, input: &'h str) -> FindMatches<'h> {
         ScannerImpl::find_iter(self.inner.clone(), input)
-    }
-
-    /// Returns the current scanner mode.
-    pub fn current_mode(&self) -> usize {
-        *self.current_mode.lock().unwrap()
-    }
-
-    /// Sets the current scanner mode.
-    ///
-    /// A parser can explicitly set the scanner mode to switch to a different set of DFAs.
-    /// Usually, the scanner mode is changed by the scanner itself based on the transitions defined
-    /// in the active scanner mode.
-    pub fn set_mode(&mut self, mode: usize) {
-        trace!("Set scanner mode to {}", mode);
-        *self.current_mode.lock().unwrap() = mode;
-    }
-
-    /// Returns the name of the scanner mode with the given index.
-    /// If the index is out of bounds, None is returned.
-    pub fn mode_name(&self, index: usize) -> Option<&str> {
-        self.inner.mode_name(index)
     }
 
     /// Logs the compiled DFAs as a Graphviz DOT file with the help of the `log` crate.
@@ -78,11 +62,36 @@ impl Scanner {
     pub fn generate_compiled_dfas_as_dot<T>(
         &self,
         modes: &[ScannerMode],
-        target_folder: T
+        target_folder: T,
     ) -> Result<()>
-        where T: AsRef<Path>
+    where
+        T: AsRef<Path>,
     {
-        self.inner.generate_compiled_dfas_as_dot(modes, target_folder)
+        self.inner
+            .generate_compiled_dfas_as_dot(modes, target_folder)
+    }
+}
+
+impl ScannerModeSwitcher for Scanner {
+    /// Returns the current scanner mode.
+    fn current_mode(&self) -> usize {
+        self.inner.current_mode()
+    }
+
+    /// Sets the current scanner mode.
+    ///
+    /// A parser can explicitly set the scanner mode to switch to a different set of DFAs.
+    /// Usually, the scanner mode is changed by the scanner itself based on the transitions defined
+    /// in the active scanner mode.
+    fn set_mode(&mut self, mode: usize) {
+        trace!("Set scanner mode to {}", mode);
+        self.inner.set_mode(mode);
+    }
+
+    /// Returns the name of the scanner mode with the given index.
+    /// If the index is out of bounds, None is returned.
+    fn mode_name(&self, index: usize) -> Option<&str> {
+        self.inner.mode_name(index)
     }
 }
 
@@ -119,13 +128,20 @@ mod tests {
         let scanner_mode = ScannerMode::new(
             "INITIAL",
             vec![(r"\r\n|\r|\n", 1), (r"(//.*(\r\n|\r|\n))", 3)],
-            vec![(1, 1), (3, 1)]
+            vec![(1, 1), (3, 1)],
         );
-        let scanner = ScannerBuilder::new().add_scanner_mode(scanner_mode).build().unwrap();
+        let scanner = ScannerBuilder::new()
+            .add_scanner_mode(scanner_mode)
+            .build()
+            .unwrap();
         assert_eq!("INITIAL", scanner.inner.scanner_modes[0].name);
         let compiled_dfa = &scanner.inner.scanner_modes[0].patterns[1].0;
 
-        compiled_dfa_render_to!(&compiled_dfa, "LineComment", scanner.inner.character_classes);
+        compiled_dfa_render_to!(
+            &compiled_dfa,
+            "LineComment",
+            scanner.inner.character_classes
+        );
     }
 
     #[test]
@@ -135,9 +151,12 @@ mod tests {
         let scanner_mode = ScannerMode::new(
             "INITIAL",
             vec![(r"\r\n|\r|\n", 1), (r"(//.*(\r\n|\r|\n))", 3)],
-            vec![(1, 1), (3, 1)]
+            vec![(1, 1), (3, 1)],
         );
-        let mut scanner = ScannerBuilder::new().add_scanner_mode(scanner_mode).build().unwrap();
+        let mut scanner = ScannerBuilder::new()
+            .add_scanner_mode(scanner_mode)
+            .build()
+            .unwrap();
         // At the beginning, the scanner mode is 0.
         assert_eq!(0, scanner.current_mode());
         assert_eq!(0, scanner.inner.current_mode());
@@ -146,17 +165,19 @@ mod tests {
         assert_eq!(1, scanner.current_mode());
         assert_eq!(1, scanner.inner.current_mode());
 
-        let find_iter = scanner.find_iter("Hello\nWorld");
-        // The creation of a find_iter resets the common scanner mode to 0.
+        let mut find_iter = scanner.find_iter("Hello\nWorld");
+        // The creation of a find_iter sets its own scanner mode to 0.
         assert_eq!(0, find_iter.current_mode());
-        assert_eq!(0, scanner.current_mode());
-        assert_eq!(0, scanner.inner.current_mode());
-        assert_eq!(0, scanner.inner.clone().current_mode());
 
-        scanner.set_mode(1);
-        assert_eq!(1, find_iter.current_mode());
         assert_eq!(1, scanner.current_mode());
         assert_eq!(1, scanner.inner.current_mode());
         assert_eq!(1, scanner.inner.clone().current_mode());
+
+        find_iter.set_mode(1);
+        assert_eq!(1, find_iter.current_mode());
+        scanner.set_mode(0);
+        assert_eq!(0, scanner.current_mode());
+        assert_eq!(0, scanner.inner.current_mode());
+        assert_eq!(0, scanner.inner.clone().current_mode());
     }
 }